chore(letters): extract letter types + remove orphan useLetterQueue hook
